re-organize script extender
