+++ Patches/ScriptExtender/Extensions/clientCreatures.cpp
