+++ Patches/ScriptExtender/Extensions/clientCreatures.h
