+++ Patches/ScriptExtender/Extensions/creatureStats.cpp
