+++ Patches/ScriptExtender/Extensions/creatureStats.h
