+++ Patches/ScriptExtender/Extensions/fileIO.cpp
