+++ Patches/ScriptExtender/Extensions/globalModifiers.cpp
