+++ Patches/ScriptExtender/Extensions/globalModifiers.h
