+++ Patches/ScriptExtender/hooks.toml
@@ -1,4 +1,3 @@
-
 # Expand the VM Commands array
 [[hooks]]
 address = 0x0054c964
