+++ Patches/ScriptExtender/manifest.toml
@@ -1,7 +1,7 @@
 [patch]
 id = "script-extender"
 name = "Lane's KotOR Script Extender"
-version = "1.0.1"
+version = "1.0.2"
 author = "Lane"
 description = "Adds additional NWScript commands."
